+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `vroom(na = )` now works when the na value is encoded in non ASCII or UTF-8 locales _and_ the file encoding is not the same as the native encoding (#233).
 
+* `vroom()` no longer truncates the last value in a file if the file contains windows newlines but no final newline (#219).
+
 * `vroom_fwf()` now verifies that the positions are valid, namely that the begin value is always less than the previous end (#217).
 
 * `vroom_lines()` gains a `locale` argument so you can control the encoding of the file (#218)